fix(release): stop a broken iOS target from blocking macOS releases
An iOS-only `module.modulemap` collision blocked every macOS release for
weeks in August 2026. The macOS app built and signed cleanly the whole
time; it was held hostage because `build-and-release` packages both and
a failing iOS archive failed the job.

The obvious fix — move iOS to its own job — is wrong here. macOS and iOS
are built from one source-coherent Signal FFI build (see "Prepare Signal
FFI XCFramework for release app"), so splitting would mean duplicating
~1300 lines of signing/FFI setup or rebuilding the FFI independently,
breaking exactly the coherence that design protects.

The defect was never the shared job; it was that iOS failure was fatal.
So iOS stays, and stops being fatal:

- the archive step is `continue-on-error: true`
- a new `ios-status` step records success/failure into the job output
  `ios_status`
- failure emits an `::error::` annotation and a job-summary block naming
  precisely what is excluded
- the iOS artifact upload and `domain-core-ios-release-evidence` both
  gate on `ios_status == 'success'`

The release degrades visibly — macOS ships, iOS is excluded, and the run
says so in three places — rather than failing wholesale or silently
publishing an incomplete set.

Also fixes the four compile errors that the modulemap collision had been
masking, and retires a duplication hazard they exposed.

`nilIfEmpty` was defined fifteen times across the mobile target with
three different behaviours: eleven tested `isEmpty` directly, three
tested the trimmed string but returned the untrimmed original, and one
returned the trimmed value. One name, three meanings, resolved by
whichever file you happened to be in.

All fifteen now route to `Models/MobileStringNilIfBlank.swift`, which
already existed for exactly this consolidation:

  nilIfBlank                      trims, returns trimmed   (pre-existing)
  nilIfEmpty                      no trim, exact-empty     (new)
  nilIfBlankPreservingWhitespace  tests trimmed, returns original (new)

Each duplicate was mapped to the variant preserving its exact prior
behaviour, so this is a pure deduplication. A first attempt introduced a
competing `nilIfBlank` with the *other* semantics, which would have
silently begun trimming 27 call sites — the same class of bug being
removed here. Caught before commit; the canonical file is extended
rather than duplicated.

Verification: `xcodebuild -scheme OpenBurnBarMobile -configuration
Release -destination generic/platform=iOS` confirms the four original
errors are gone. A second layer of pre-existing breakage remains in the
iOS target (two non-exhaustive switches, a main-actor isolation
violation, a `()`->`AuthCredential` mismatch, and a type-check timeout in
RootTabView) — deliberately NOT fixed here. That is the breakage this
change makes survivable rather than release-blocking.

Contract suites: 72 pass / 0 fail across native-release workflows,
domain-core workflows, control-plane, reachability, and xcframework
collision.

The manifest diff also carries three digests that were already stale on
main before this branch (#2343, #2344, #2347 each edited a pinned script
without regenerating). #2348 fixes the same three; the values are
identical, so whichever lands first makes the other a no-op.

Runbook gains a section on this design, plus corrections from a peer
session narrowing the manifest rule: the pinned set spans `scripts/`
broadly, not just workflows, and the reflexive check is to run `--write`
and confirm the diff names only files you edited.

+++ OpenBurnBarMobile/Views/Pulse/PulseWindowMetrics.swift
@@ -21,7 +21,9 @@ enum PulseWindowMetricBuilder {
             rollups: Dictionary(uniqueKeysWithValues: rollupTotals.map { key, value in
                 (key.rawValue, MobilePulseRollupTotals(
                     requests: value.requests,
-                    tokens: value.tokens,
+                    // RollupTotals.tokens is Int; the policy model widens to
+                    // Int64. Widening is total, so no clamping is needed.
+                    tokens: Int64(value.tokens),
                     costUsd: value.costUsd
                 ))
             }),
